# Fix recursive perf project build and add memory diagnoser