--- a/performance/SqlOutputBindingPerformance.cs
+++ b/performance/SqlOutputBindingPerformance.cs
@@ -13,6 +13,7 @@ using BenchmarkDotNet.Attributes;
 
 namespace Microsoft.Azure.WebJobs.Extensions.Sql.Performance
 {
+    [MemoryDiagnoser]
     public class SqlOutputBindingPerformance : IntegrationTestBase
     {
         [GlobalSetup]